pyproject.toml
[project]
name = "photorank"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "dspy>=2.6.27",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]
